requirements.txt
streamlit
embedchain
python-dotenv 